requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
lxml>=4.9.0
imagehash>=4.3.0
